refactor(app): isolate timeline virtualizer
