+++ packages/app/src/pages/session/timeline/message-timeline.tsx
@@ -1,17 +1,5 @@
-import {
-  createEffect,
-  createMemo,
-  createSignal,
-  For,
-  on,
-  onCleanup,
-  onMount,
-  Show,
-  type Accessor,
-  type JSX,
-} from "solid-js"
+import { createEffect, createMemo, createSignal, on, Show, type Accessor, type JSX } from "solid-js"
 import { createStore } from "solid-js/store"
-import { createVirtualizer, defaultRangeExtractor, elementScroll, type VirtualItem } from "@tanstack/solid-virtual"
 import type { SessionUserActions } from "@opencode-ai/session-ui/actions"
 import { DiffChanges } from "@opencode-ai/ui/diff-changes"
 import { Icon } from "@opencode-ai/ui/icon"
@@ -21,61 +9,23 @@ import { Menu } from "@opencode-ai/ui/menu"
 import { Tooltip } from "@opencode-ai/ui/tooltip"
 import { ProjectAvatar } from "@opencode-ai/ui/project-avatar"
 import { Button } from "@opencode-ai/ui/button"
-import { isScrollKeyTarget, scrollKey, scrollKeyOwner, ScrollView } from "@opencode-ai/ui/scroll-view"
 import type { Project } from "@/types"
 import { getFilename } from "@opencode-ai/util/path"
 import { Popover as KobaltePopover } from "@kobalte/core/popover"
-import { shouldMarkBoundaryGesture, normalizeWheelDelta } from "@/pages/session/message-gesture"
 import { SessionContextUsage } from "@/components/session-context-usage"
 import { useLanguage } from "@/context/language"
 import { useData } from "@/context/server"
 import { useWorkspaceLocation } from "@/context/location"
-import { scheduleConnectedMeasure } from "./measure"
-import { observeElementOffsetReconnectAware } from "./observe-element-offset"
-import { Timeline, TimelineRow } from "@opencode-ai/session-ui/timeline/projection"
+import { Timeline } from "@opencode-ai/session-ui/timeline/projection"
 import { createSessionTimelineRowRenderer } from "@opencode-ai/session-ui/timeline/row"
-import { filterVirtualIndexes } from "./virtual-items"
 import { createTimelineController, type TimelineController, type TimelineSessionSource } from "./controller"
+import { createTimelineVirtualizer } from "./virtualizer"
 import { containsDirectory, isWorkspaceDirectory, workspaceDirectories } from "@/utils/workspace"
 import { SessionWorkspaceMenu } from "@/components/session-workspace-menu"
 import { getProjectAvatarVariant } from "@/context/layout"
 import { displayName, getProjectAvatarSource } from "@/pages/layout/helpers"
 import type { SessionMessageUser } from "@opencode-ai/client/promise"
 import { parseCommentNote, readPromptPresentation } from "@/utils/comment-note"
-
-const timelineFallbackItemSize = 60
-const timelineCache = new Map<string, { measurements: VirtualItem[]; toolOpen: Record<string, boolean | undefined> }>()
-
-const boundaryTarget = (root: HTMLElement, target: EventTarget | null) => {
-  const current = target instanceof Element ? target : undefined
-  const nested = current?.closest("[data-scrollable]")
-  if (!nested || nested === root) return root
-  if (!(nested instanceof HTMLElement)) return root
-  return nested
-}
-
-const markBoundaryGesture = (input: {
-  root: HTMLDivElement
-  target: EventTarget | null
-  delta: number
-  onMarkScrollGesture: (target?: EventTarget | null) => void
-}) => {
-  const target = boundaryTarget(input.root, input.target)
-  if (target === input.root) {
-    input.onMarkScrollGesture(input.root)
-    return
-  }
-  if (
-    shouldMarkBoundaryGesture({
-      delta: input.delta,
-      scrollTop: target.scrollTop,
-      scrollHeight: target.scrollHeight,
-      clientHeight: target.clientHeight,
-    })
-  ) {
-    input.onMarkScrollGesture(input.root)
-  }
-}
 
 function WorkspaceMoveAction(props: {
   variant: "inline" | "panel"
@@ -277,18 +227,9 @@ function MessageTimelineView(
     pending: TimelineController["pending"]
   },
 ) {
-  let touchGesture: number | undefined
   const language = useLanguage()
   const data = useData()
   const sdk = useWorkspaceLocation()
-  const shouldAnchorBottom = createMemo(() => props.shouldAnchorBottom)
-  const hasScrollGesture = createMemo(() => props.hasScrollGesture)
-  const ownerSessionKey = props.data.sessionKey()
-  const cached = timelineCache.get(ownerSessionKey)
-  const initialMeasurements = cached?.measurements
-  const coldBottomMount = !initialMeasurements?.length && shouldAnchorBottom()
-
-  const [listRoot, setListRoot] = createSignal<HTMLDivElement>()
   const sessionID = props.data.sessionID
   const sessionStatus = props.data.status
   const titleLabel = props.data.titleLabel
@@ -328,215 +269,28 @@ function MessageTimelineView(
   )
   const turnPadding = () => "px-4 md:px-5"
   const showHeader = createMemo(() => props.data.showHeader() || workspaceSession())
-  const activeMessageID = projection.activeMessageID
   const messageByID = projection.messageByID
-  const messageLastRowIndex = projection.messageLastRowIndex
-  const messageRowIndex = projection.messageRowIndex
-  const timelineRowByKey = projection.rowByKey
-  const timelineRows = projection.rows
-  let prependAnchor: { key: string; offset: number } | undefined
-  let prependAnchorFrame: number | undefined
-  let prependLoading = false
-  const clearPrependAnchor = () => {
-    prependLoading = false
-    prependAnchor = undefined
-    if (prependAnchorFrame === undefined) return
-    cancelAnimationFrame(prependAnchorFrame)
-    prependAnchorFrame = undefined
-  }
-  const capturePrependAnchor = () => {
-    prependLoading = true
-    updatePrependAnchor()
-  }
-  const updatePrependAnchor = () => {
-    const root = listRoot()
-    if (!root) return
-    const view = root.getBoundingClientRect()
-    const anchor = [...root.querySelectorAll<HTMLElement>("[data-timeline-key]")]
-      .map((element) => ({ element, rect: element.getBoundingClientRect() }))
-      .filter((item) => item.rect.bottom > view.top && item.rect.top < view.bottom)
-      .sort((a, b) => a.rect.top - b.rect.top)[0]
-    if (!anchor) return
-    if (!anchor.element.dataset.timelineKey) return
-    prependAnchor = { key: anchor.element.dataset.timelineKey, offset: anchor.rect.top - view.top }
-  }
-  const restorePrependAnchor = (done: boolean) => {
-    if (done) prependLoading = false
-    applyPrependAnchor()
-  }
-  const applyPrependAnchor = () => {
-    const root = listRoot()
-    if (!root || !prependAnchor) return
-    if (prependAnchorFrame !== undefined) cancelAnimationFrame(prependAnchorFrame)
-    let frames = 0
-    let stable = 0
-    const apply = () => {
-      prependAnchorFrame = undefined
-      const anchor = prependAnchor
-      if (!anchor) return
-      const element = root.querySelector<HTMLElement>(`[data-timeline-key="${CSS.escape(anchor.key)}"]`)
-      const delta = element
-        ? element.getBoundingClientRect().top - root.getBoundingClientRect().top - anchor.offset
-        : undefined
-      if (delta !== undefined && Math.abs(delta) > 0.5) {
-        root.scrollTop += delta
-        stable = 0
-      } else {
-        stable += 1
-      }
-      frames += 1
-      if (stable >= 30 || frames >= 180) {
-        if (!prependLoading) prependAnchor = undefined
-        return
-      }
-      prependAnchorFrame = requestAnimationFrame(apply)
-    }
-    prependAnchorFrame = requestAnimationFrame(apply)
-  }
-
-  const [toolOpen, setToolOpen] = createStore<Record<string, boolean | undefined>>(cached?.toolOpen ?? {})
-  const [renderOverscan, setRenderOverscan] = createSignal(initialMeasurements?.length || coldBottomMount ? 6 : 20)
-  let resizePinnedIndexes: number[] = []
-  let resizePinFrame: number | undefined
-  let virtualContent: HTMLDivElement | undefined
-  const virtualizer = createVirtualizer<HTMLDivElement, HTMLDivElement>({
-    get count() {
-      return timelineRows().length
-    },
-    getScrollElement: () => listRoot() ?? null,
-    observeElementOffset: observeElementOffsetReconnectAware,
-    initialOffset: () => (shouldAnchorBottom() ? Number.MAX_SAFE_INTEGER : 0),
-    initialMeasurementsCache: initialMeasurements,
-    estimateSize: () => timelineFallbackItemSize,
-    scrollToFn: (offset, options, instance) => {
-      // Expose the computed range before core writes an anchor correction so the browser does not clamp it to the old height.
-      if (virtualContent) virtualContent.style.height = `${instance.getTotalSize()}px`
-      elementScroll(offset, options, instance)
-    },
-    get getItemKey() {
-      const rows = timelineRows()
-      return (index: number) => {
-        const row = rows[index]
-        // ResizeObserver can report a removed element after its row has left the projection.
-        if (!row) return `removed:${index}`
-        return TimelineRow.key(row)
-      }
-    },
-    anchorTo: "end",
-    followOnAppend: true,
-    scrollEndThreshold: 80,
-    get scrollMargin() {
-      return showHeader() ? 64 : 0
-    },
-    overscan: 50,
-    paddingEnd: 64,
-    rangeExtractor: (range) => {
-      const id = activeMessageID()
-      const active = id ? (messageLastRowIndex().get(id) ?? -1) : -1
-      const indexes = defaultRangeExtractor({ ...range, overscan: renderOverscan() })
-      return filterVirtualIndexes(
-        [...new Set([...resizePinnedIndexes, ...indexes, ...(active < 0 ? [] : [active])])].sort((a, b) => a - b),
-        range.count,
-      )
-    },
+  const virtualized = createTimelineVirtualizer({
+    sessionKey: props.data.sessionKey,
+    projection,
+    showHeader,
+    shouldAnchorBottom: () => props.shouldAnchorBottom,
+    hasScrollGesture: () => props.hasScrollGesture,
+    scroll: () => props.scroll,
+    onResumeScroll: props.onResumeScroll,
+    setScrollRef: props.setScrollRef,
+    setContentRef: props.setContentRef,
+    onScheduleScrollState: props.onScheduleScrollState,
+    onAutoScrollHandleScroll: props.onAutoScrollHandleScroll,
+    onAutoScrollInteraction: props.onAutoScrollInteraction,
+    onMarkScrollGesture: props.onMarkScrollGesture,
+    onUserScroll: props.onUserScroll,
+    onHistoryScroll: props.onHistoryScroll,
+    setRevealMessage: props.setRevealMessage,
+    setScrollToEnd: props.setScrollToEnd,
+    setHistoryAnchor: props.setHistoryAnchor,
   })
-  const resizeItem = virtualizer.resizeItem
-  let resizeAnchorScheduled = false
-  const anchorResizedBottom = () => {
-    if (resizeAnchorScheduled || hasScrollGesture()) return
-    resizeAnchorScheduled = true
-    queueMicrotask(() => {
-      resizeAnchorScheduled = false
-      if (!shouldAnchorBottom() || hasScrollGesture()) return
-      virtualizer.scrollToEnd()
-    })
-  }
-  virtualizer.resizeItem = (index, size) => {
-    const item = virtualizer.measurementsCache[index]
-    const previous = item ? (virtualizer.itemSizeCache.get(item.key) ?? item.size) : undefined
-    const root = listRoot()
-    if (root && previous !== undefined && Math.abs(size - previous) > root.clientHeight) {
-      const view = root.getBoundingClientRect()
-      resizePinnedIndexes = [...root.querySelectorAll<HTMLElement>("[data-index]")]
-        .filter((element) => {
-          const rect = element.getBoundingClientRect()
-          return rect.bottom > view.top && rect.top < view.bottom
-        })
-        .map((element) => Number(element.dataset.index))
-      if (resizePinFrame !== undefined) cancelAnimationFrame(resizePinFrame)
-      resizePinFrame = requestAnimationFrame(() => {
-        resizePinFrame = requestAnimationFrame(() => {
-          resizePinFrame = undefined
-          resizePinnedIndexes = []
-        })
-      })
-    }
-    resizeItem(index, size)
-    if (root && shouldAnchorBottom()) anchorResizedBottom()
-  }
-  virtualizer.shouldAdjustScrollPositionOnItemSizeChange = (item) => {
-    if (shouldAnchorBottom()) return false
-    const first = virtualizer.range?.startIndex
-    return first !== undefined && item.index < first
-  }
-  const virtualItemByKey = createMemo(
-    () => new Map(virtualizer.getVirtualItems().map((item) => [item.key, item] as const)),
-  )
-  const virtualRowKeys = createMemo(() => virtualizer.getVirtualItems().map((item) => String(item.key)))
-  createEffect(() => {
-    props.setRevealMessage?.((id) => {
-      const index = messageRowIndex().get(id)
-      if (index === undefined) return
-      virtualizer.scrollToIndex(index, { align: "center" })
-    })
-    props.setScrollToEnd?.(() => virtualizer.scrollToEnd())
-    props.setHistoryAnchor?.({ capture: capturePrependAnchor, restore: restorePrependAnchor })
-  })
-
-  let overscanFrame: number | undefined
-  onMount(() => {
-    overscanFrame = requestAnimationFrame(() => {
-      if (shouldAnchorBottom()) virtualizer.scrollToEnd()
-      overscanFrame = requestAnimationFrame(() => {
-        overscanFrame = undefined
-        if (renderOverscan() < 20) setRenderOverscan(20)
-        if (shouldAnchorBottom()) virtualizer.scrollToEnd()
-      })
-    })
-  })
-
-  const maybeAnchorBottom = () => {
-    if (timelineRows().length === 0) return
-    if (!shouldAnchorBottom() || hasScrollGesture()) return
-    if (resizePinFrame !== undefined) cancelAnimationFrame(resizePinFrame)
-    clearPrependAnchor()
-    if (prependAnchorFrame !== undefined) cancelAnimationFrame(prependAnchorFrame)
-    virtualizer.scrollToEnd()
-  }
-
-  let measuredSessionKey = props.data.sessionKey()
-  createEffect(() => {
-    const key = props.data.sessionKey()
-    timelineRows().length
-    if (measuredSessionKey !== key) {
-      measuredSessionKey = key
-      virtualizer.measure()
-    }
-    maybeAnchorBottom()
-  })
-
-  onCleanup(() => {
-    clearPrependAnchor()
-    timelineCache.delete(ownerSessionKey)
-    timelineCache.set(ownerSessionKey, { measurements: virtualizer.takeSnapshot(), toolOpen: { ...toolOpen } })
-    while (timelineCache.size > 16) timelineCache.delete(timelineCache.keys().next().value!)
-    if (resizePinFrame !== undefined) cancelAnimationFrame(resizePinFrame)
-    if (overscanFrame !== undefined) cancelAnimationFrame(overscanFrame)
-    props.setRevealMessage?.(() => {})
-    props.setScrollToEnd?.(() => {})
-    props.setHistoryAnchor?.({ capture: () => {}, restore: () => {} })
-  })
-
+  const VirtualizedTimeline = virtualized.View
   const [title, setTitle] = createStore({
     draft: "",
     editing: false,
@@ -551,83 +305,6 @@ function MessageTimelineView(
     dismiss: null as "escape" | "outside" | null,
   })
   let more: HTMLButtonElement | undefined
-
-  const bindListRoot = (root: HTMLDivElement) => {
-    if (root === listRoot()) return
-    setListRoot(root)
-    props.setScrollRef(root)
-  }
-
-  const handleListWheel = (event: WheelEvent & { currentTarget: HTMLDivElement }) => {
-    if (!prependLoading) clearPrependAnchor()
-    const root = event.currentTarget
-    const delta = normalizeWheelDelta({
-      deltaY: event.deltaY,
-      deltaMode: event.deltaMode,
-      rootHeight: root.clientHeight,
-    })
-    if (!delta) return
-    markBoundaryGesture({ root, target: event.target, delta, onMarkScrollGesture: props.onMarkScrollGesture })
-  }
-
-  const handleListTouchStart = (event: TouchEvent) => {
-    if (!prependLoading) clearPrependAnchor()
-    touchGesture = event.touches[0]?.clientY
-  }
-
-  const handleListTouchMove = (event: TouchEvent & { currentTarget: HTMLDivElement }) => {
-    const next = event.touches[0]?.clientY
-    const prev = touchGesture
-    touchGesture = next
-    if (next === undefined || prev === undefined) return
-
-    const delta = prev - next
-    if (!delta) return
-
-    markBoundaryGesture({
-      root: event.currentTarget,
-      target: event.target,
-      delta,
-      onMarkScrollGesture: props.onMarkScrollGesture,
-    })
-  }
-
-  const handleListTouchEnd = () => {
-    touchGesture = undefined
-  }
-
-  const handleListPointerDown = (event: PointerEvent & { currentTarget: HTMLDivElement }) => {
-    if (!prependLoading) clearPrependAnchor()
-    props.onMarkScrollGesture(event.target)
-  }
-
-  const handleListPointerMove = (event: PointerEvent) => {
-    if (event.buttons !== 1) return
-    props.onMarkScrollGesture(event.target)
-  }
-
-  const handleListKeyDown = (event: KeyboardEvent & { currentTarget: HTMLDivElement }) => {
-    const key = scrollKey(event)
-    if (!key) return
-    if (!isScrollKeyTarget(event.target, key)) return
-    if (scrollKeyOwner(event.currentTarget, event.target, key) !== event.currentTarget) return
-    if (!prependLoading) clearPrependAnchor()
-    props.onMarkScrollGesture(event.currentTarget)
-  }
-
-  const handleListScroll = (event: Event & { currentTarget: HTMLDivElement }) => {
-    if (prependLoading) updatePrependAnchor()
-    props.onScheduleScrollState(event.currentTarget)
-    props.onHistoryScroll()
-    if (!props.hasScrollGesture) return
-    props.onUserScroll()
-    props.onAutoScrollHandleScroll()
-    props.onMarkScrollGesture(event.currentTarget)
-  }
-
-  onCleanup(() => {
-    props.setScrollRef(undefined)
-  })
 
   const selectShareUrlText: JSX.EventHandler<HTMLDivElement, MouseEvent> = (event) => {
     const selection = window.getSelection()
@@ -688,479 +365,337 @@ function MessageTimelineView(
     showReasoningSummaries: props.data.showReasoningSummaries,
     shellToolDefaultOpen: props.data.shellToolPartsExpanded,
     editToolDefaultOpen: props.data.editToolPartsExpanded,
-    disclosure: {
-      value: (key) => toolOpen[key],
-      set: (key, open) => setToolOpen(key, open),
-    },
+    disclosure: virtualized.disclosure,
     centered: () => props.centered,
     padding: turnPadding,
     anchor: props.anchor,
   })
 
-  function TimelineRowView(props: { row: TimelineRow.TimelineRow; onSizeChange?: () => void }) {
-    return <rowRenderer.Row row={() => props.row} onSizeChange={props.onSizeChange} />
-  }
-
-  function VirtualTimelineRow(props: { rowKey: string }) {
-    let element: HTMLDivElement
-    const initialItem = virtualItemByKey().get(props.rowKey)!
-    const initialRow = timelineRowByKey().get(props.rowKey)!
-    const item = createMemo(() => virtualItemByKey().get(props.rowKey) ?? initialItem)
-    const row = createMemo(() => timelineRowByKey().get(props.rowKey) ?? timelineRows()[item().index] ?? initialRow)
-    const tool = () => {
-      const value = row()
-      if (value._tag !== "AssistantPart" || value.group.type !== "part") return undefined
-      const content = Timeline.resolveContent(messageByID().get(value.group.ref.messageID), value.group.ref.partID)
-      if (content?.type === "tool") return content
-      return undefined
-    }
-    const asyncFile = () => ["edit", "write", "patch"].includes(tool()?.name ?? "")
-    const [ready, setReady] = createSignal(initialItem.size <= timelineFallbackItemSize || !asyncFile())
-    let contentMeasureFrame: number | undefined
-
-    onMount(() => virtualizer.measureElement(element))
-
-    createEffect(
-      on(
-        () => item().index,
-        () => {
-          virtualizer.measureElement(element)
-        },
-        { defer: true },
-      ),
-    )
-
-    onCleanup(() => {
-      if (contentMeasureFrame !== undefined) cancelAnimationFrame(contentMeasureFrame)
-      // Solid runs cleanup before it disconnects the row, so defer TanStack's null-ref cleanup.
-      queueMicrotask(() => virtualizer.measureElement(null))
-    })
-
-    return (
-      <div
-        data-timeline-key={props.rowKey}
-        style={{
-          position: "absolute",
-          top: `${item().start - (showHeader() ? 64 : 0)}px`,
-          left: "0",
-          width: "100%",
-          height: `${item().size}px`,
-          overflow: "clip",
-          // Rounded virtual measurements can otherwise clip a framed row's outer paint.
-          "overflow-clip-margin": row()._tag === "TurnGap" ? undefined : "0.5px",
-        }}
-      >
-        <div
-          ref={(value) => {
-            element = value
-          }}
-          data-index={item().index}
-          style={{ "min-height": ready() ? undefined : `${initialItem.size}px` }}
-        >
-          <TimelineRowView
-            row={row()}
-            onSizeChange={() => {
-              setReady(true)
-              if (contentMeasureFrame !== undefined) cancelAnimationFrame(contentMeasureFrame)
-              contentMeasureFrame = scheduleConnectedMeasure(element, virtualizer.measureElement)
-            }}
-          />
-        </div>
-      </div>
-    )
-  }
-
   return (
-    <div class="relative w-full h-full min-w-0" data-workspace-session={workspaceSession() ? "" : undefined}>
-      <div
-        class="absolute left-1/2 -translate-x-1/2 z-[60] pointer-events-none transition-all duration-200 ease-out"
-        classList={{
-          "bottom-8": true,
-          "opacity-100 translate-y-0 scale-100": props.scroll.overflow && props.scroll.jump,
-          "opacity-0 translate-y-2 pointer-events-none": !props.scroll.overflow || !props.scroll.jump,
-          "scale-[0.8]": !props.scroll.overflow || !props.scroll.jump,
-        }}
-      >
-        <button
-          type="button"
-          aria-label={language.t("session.messages.jumpToLatest")}
-          class="pointer-events-auto flex items-center justify-center w-8 h-7 px-2 py-1.5 rounded-lg border-none cursor-pointer text-v2-text-text-base backdrop-blur-[2px]"
-          style={{
-            background: "color-mix(in srgb, var(--v2-background-bg-base) 92%, transparent)",
-            "box-shadow": "var(--v2-elevation-raised), 0px 2px 8px var(--v2-background-bg-base)",
-          }}
-          onClick={props.onResumeScroll}
+    <VirtualizedTimeline
+      workspaceSession={workspaceSession}
+      deferred={(row) => {
+        if (row._tag !== "AssistantPart" || row.group.type !== "part") return false
+        const content = Timeline.resolveContent(messageByID().get(row.group.ref.messageID), row.group.ref.partID)
+        return content?.type === "tool" && ["edit", "write", "patch"].includes(content.name)
+      }}
+      renderRow={(row, onSizeChange) => <rowRenderer.Row row={row} onSizeChange={onSizeChange} />}
+      header={
+        <div
+          data-session-title
+          class="sticky top-0 z-30 bg-[linear-gradient(to_bottom,var(--v2-background-bg-base)_48px,transparent)] w-full pb-4 pr-3 pl-2.5"
         >
-          <svg width="16" height="16" viewBox="0 0 16 16" fill="none" aria-hidden="true">
-            <path
-              d="M12.3333 8.66665L8 13L3.66667 8.66665M8 12.6667V2.83332"
-              stroke="currentColor"
-              stroke-linecap="square"
-            />
-          </svg>
-        </button>
-      </div>
-      <ScrollView
-        viewportRef={bindListRoot}
-        onWheel={handleListWheel}
-        onTouchStart={handleListTouchStart}
-        onTouchMove={handleListTouchMove}
-        onTouchEnd={handleListTouchEnd}
-        onTouchCancel={handleListTouchEnd}
-        onPointerDown={handleListPointerDown}
-        onPointerMove={handleListPointerMove}
-        onKeyDown={handleListKeyDown}
-        onScroll={handleListScroll}
-        onClick={props.onAutoScrollInteraction}
-        class="relative min-w-0 w-full h-full"
-        style={{
-          "--sticky-accordion-top": showHeader() ? "48px" : "0px",
-        }}
-      >
-        <Show when={showHeader()}>
-          <div
-            data-session-title
-            class="sticky top-0 z-30 bg-[linear-gradient(to_bottom,var(--v2-background-bg-base)_48px,transparent)] w-full pb-4 pr-3 pl-2.5"
-          >
-            <div class="h-12 w-full flex items-center justify-between gap-2">
-              <div class="flex items-center gap-1 min-w-0 flex-1">
-                <div class="flex items-center min-w-0 flex-1 w-full">
+          <div class="h-12 w-full flex items-center justify-between gap-2">
+            <div class="flex items-center gap-1 min-w-0 flex-1">
+              <div class="flex items-center min-w-0 flex-1 w-full">
+                <Show
+                  when={workspaceSession()}
+                  fallback={
+                    <span class="flex size-6 shrink-0 items-center justify-center text-v2-icon-icon-muted">
+                      <Icon name="monitor" />
+                    </span>
+                  }
+                >
+                  <Tooltip
+                    placement="bottom-start"
+                    value={sessionDirectory()}
+                    contentClass="max-w-[calc(100vw-32px)] break-all"
+                  >
+                    <span
+                      tabIndex={0}
+                      aria-label={sessionDirectory()}
+                      class="flex size-6 shrink-0 items-center justify-center text-v2-icon-icon-accent"
+                    >
+                      <Icon name="workspace-isolated" />
+                    </span>
+                  </Tooltip>
+                </Show>
+                <Show when={parentID()}>
+                  <button
+                    type="button"
+                    data-slot="session-title-parent"
+                    class="min-w-0 max-w-[40%] truncate pl-2 text-[13px] font-[530] leading-4 tracking-[-0.04px] text-v2-text-text-faint transition-colors hover:text-v2-text-text-muted"
+                    onClick={props.action.navigateParent}
+                  >
+                    {parentTitle()}
+                  </button>
+                  <span
+                    data-slot="session-title-separator"
+                    class="-translate-y-[0.5px] pl-2 pr-1 text-[11px] font-medium text-v2-text-text-faint"
+                    aria-hidden="true"
+                  >
+                    /
+                  </span>
+                </Show>
+                <Show when={childTitle() || title.editing}>
                   <Show
-                    when={workspaceSession()}
+                    when={title.editing}
                     fallback={
-                      <span class="flex size-6 shrink-0 items-center justify-center text-v2-icon-icon-muted">
-                        <Icon name="monitor" />
-                      </span>
+                      <h1
+                        data-slot="session-title-child"
+                        class="truncate text-[13px] font-[530] leading-4 tracking-[-0.04px] text-v2-text-text-base w-fit rounded-[6px] px-2 py-1 hover:bg-v2-overlay-simple-overlay-hover"
+                        onClick={openTitleEditor}
+                      >
+                        {childTitle()}
+                      </h1>
                     }
                   >
-                    <Tooltip
-                      placement="bottom-start"
-                      value={sessionDirectory()}
-                      contentClass="max-w-[calc(100vw-32px)] break-all"
-                    >
-                      <span
-                        tabIndex={0}
-                        aria-label={sessionDirectory()}
-                        class="flex size-6 shrink-0 items-center justify-center text-v2-icon-icon-accent"
-                      >
-                        <Icon name="workspace-isolated" />
-                      </span>
-                    </Tooltip>
+                    <InlineInput
+                      ref={(el) => {
+                        titleRef = el
+                      }}
+                      data-slot="session-title-child"
+                      dir="auto"
+                      value={title.draft}
+                      disabled={props.pending.rename()}
+                      class="block text-[13px] font-[530] leading-4 tracking-[-0.04px] text-v2-text-text-base field-sizing-content self-start rounded-[6px] px-2 py-1"
+                      style={{
+                        "--inline-input-shadow": "none",
+                        "text-align": "start",
+                      }}
+                      onInput={(event) => setTitle("draft", event.currentTarget.value)}
+                      onKeyDown={(event) => {
+                        event.stopPropagation()
+                        if (event.key === "Enter") {
+                          event.preventDefault()
+                          void saveTitleEditor()
+                          return
+                        }
+                        if (event.key === "Escape") {
+                          event.preventDefault()
+                          closeTitleEditor()
+                        }
+                      }}
+                      onBlur={closeTitleEditor}
+                    />
                   </Show>
-                  <Show when={parentID()}>
-                    <button
-                      type="button"
-                      data-slot="session-title-parent"
-                      class="min-w-0 max-w-[40%] truncate pl-2 text-[13px] font-[530] leading-4 tracking-[-0.04px] text-v2-text-text-faint transition-colors hover:text-v2-text-text-muted"
-                      onClick={props.action.navigateParent}
-                    >
-                      {parentTitle()}
-                    </button>
-                    <span
-                      data-slot="session-title-separator"
-                      class="-translate-y-[0.5px] pl-2 pr-1 text-[11px] font-medium text-v2-text-text-faint"
-                      aria-hidden="true"
-                    >
-                      /
-                    </span>
-                  </Show>
-                  <Show when={childTitle() || title.editing}>
-                    <Show
-                      when={title.editing}
-                      fallback={
-                        <h1
-                          data-slot="session-title-child"
-                          class="truncate text-[13px] font-[530] leading-4 tracking-[-0.04px] text-v2-text-text-base w-fit rounded-[6px] px-2 py-1 hover:bg-v2-overlay-simple-overlay-hover"
-                          onClick={openTitleEditor}
-                        >
-                          {childTitle()}
-                        </h1>
-                      }
-                    >
-                      <InlineInput
-                        ref={(el) => {
-                          titleRef = el
-                        }}
-                        data-slot="session-title-child"
-                        dir="auto"
-                        value={title.draft}
-                        disabled={props.pending.rename()}
-                        class="block text-[13px] font-[530] leading-4 tracking-[-0.04px] text-v2-text-text-base field-sizing-content self-start rounded-[6px] px-2 py-1"
-                        style={{
-                          "--inline-input-shadow": "none",
-                          "text-align": "start",
-                        }}
-                        onInput={(event) => setTitle("draft", event.currentTarget.value)}
-                        onKeyDown={(event) => {
-                          event.stopPropagation()
-                          if (event.key === "Enter") {
-                            event.preventDefault()
-                            void saveTitleEditor()
-                            return
-                          }
-                          if (event.key === "Escape") {
-                            event.preventDefault()
-                            closeTitleEditor()
-                          }
-                        }}
-                        onBlur={closeTitleEditor}
-                      />
-                    </Show>
-                  </Show>
-                </div>
+                </Show>
               </div>
-              <Show when={sessionID()} keyed>
-                {(id) => (
-                  <div class="shrink-0 flex items-center gap-2">
-                    <SessionContextUsage placement="bottom" />
-                    <Show when={!parentID() && project()}>
-                      {(project) => (
-                        <KobaltePopover
-                          open={summaryOpen()}
-                          placement="bottom-end"
-                          gutter={6}
-                          onOpenChange={setSummary}
-                        >
-                          <KobaltePopover.Trigger
-                            as={IconButton}
-                            icon={<Icon name="window-analytics" />}
-                            variant="ghost-muted"
-                            size="large"
-                            state={summaryOpen() ? "pressed" : undefined}
-                            aria-label={language.t("session.summary.title")}
-                            aria-expanded={summaryOpen()}
-                          />
-                          <KobaltePopover.Portal>
-                            <KobaltePopover.Content class="z-50 border-0 bg-transparent p-0 outline-none">
-                              <SessionSummaryPanel
-                                project={project()}
-                                directory={sessionDirectory()}
-                                local={!workspaceSession()}
-                                branch={data.location.vcs.info({ directory: sdk().directory })?.branch.current}
-                                baseBranch={data.location.vcs.info({ directory: project().worktree })?.branch.current}
-                                diffs={sessionDiffs()}
-                                sessionID={id}
-                                moveEligible={props.workspaceMoveEligible}
-                                moveDismissed={workspaceSuggestionDismissed()}
-                                onMoveDismiss={() => setWorkspaceSuggestionDismissed(true)}
-                                onReview={() => {
-                                  setSummary(false)
-                                  props.onReview()
-                                }}
-                              />
-                            </KobaltePopover.Content>
-                          </KobaltePopover.Portal>
-                        </KobaltePopover>
-                      )}
-                    </Show>
-                    <Show when={!parentID()}>
-                      <Menu
-                        gutter={6}
-                        placement="bottom-end"
-                        open={title.menuOpen}
-                        onOpenChange={(open) => {
-                          setTitle("menuOpen", open)
-                          if (open) return
-                        }}
-                      >
-                        <Menu.Trigger
+            </div>
+            <Show when={sessionID()} keyed>
+              {(id) => (
+                <div class="shrink-0 flex items-center gap-2">
+                  <SessionContextUsage placement="bottom" />
+                  <Show when={!parentID() && project()}>
+                    {(project) => (
+                      <KobaltePopover open={summaryOpen()} placement="bottom-end" gutter={6} onOpenChange={setSummary}>
+                        <KobaltePopover.Trigger
                           as={IconButton}
-                          icon={<Icon name="outline-dots" />}
+                          icon={<Icon name="window-analytics" />}
                           variant="ghost-muted"
                           size="large"
-                          state={share.open || title.pendingShare ? "pressed" : undefined}
-                          aria-label={language.t("common.moreOptions")}
-                          aria-expanded={title.menuOpen || share.open || title.pendingShare}
-                          ref={(el: HTMLButtonElement) => {
-                            more = el
-                          }}
+                          state={summaryOpen() ? "pressed" : undefined}
+                          aria-label={language.t("session.summary.title")}
+                          aria-expanded={summaryOpen()}
                         />
-                        <Menu.Portal>
-                          <Menu.Content
-                            style={{ width: "120px", "min-width": "120px" }}
-                            onCloseAutoFocus={(event) => {
-                              if (title.pendingRename) {
-                                event.preventDefault()
-                                setTitle("pendingRename", false)
-                                openTitleEditor()
-                                return
-                              }
-                              if (title.pendingShare) {
-                                event.preventDefault()
-                                requestAnimationFrame(() => {
-                                  setShare({ open: true, dismiss: null })
-                                  setTitle("pendingShare", false)
-                                })
-                              }
-                            }}
-                          >
-                            <Menu.Item
-                              onSelect={() => {
-                                setTitle("pendingRename", true)
-                                setTitle("menuOpen", false)
-                              }}
-                            >
-                              {language.t("common.rename")}
-                            </Menu.Item>
-                            <Show when={shareEnabled()}>
-                              <Menu.Item
-                                onSelect={() => {
-                                  setTitle({ pendingShare: true, menuOpen: false })
-                                }}
-                              >
-                                {language.t("session.share.action.share")}...
-                              </Menu.Item>
-                            </Show>
-                            <Menu.Item onSelect={() => void props.action.export(id)}>
-                              {language.t("common.export")}...
-                            </Menu.Item>
-                            {/* TODO: Need a V2 session archive API. */}
-                            <Menu.Separator />
-                            <Menu.Item onSelect={() => props.action.showDelete(id)}>
-                              {language.t("common.delete")}...
-                            </Menu.Item>
-                          </Menu.Content>
-                        </Menu.Portal>
-                      </Menu>
-
-                      <KobaltePopover
-                        open={share.open}
-                        anchorRef={() => more}
-                        placement="bottom-end"
-                        gutter={6}
-                        modal={false}
-                        onOpenChange={(open) => {
-                          if (open) setShare("dismiss", null)
-                          setShare("open", open)
-                        }}
-                      >
                         <KobaltePopover.Portal>
-                          <KobaltePopover.Content
-                            data-component="popover-content"
-                            class="flex w-80 max-w-none flex-col items-start gap-3 rounded-[10px] border-0 bg-v2-background-bg-layer-01 p-3 shadow-[var(--v2-elevation-floating)]"
-                            style={{ "min-width": "320px" }}
-                            onEscapeKeyDown={(event) => {
-                              setShare({ dismiss: "escape", open: false })
-                              event.preventDefault()
-                              event.stopPropagation()
-                            }}
-                            onPointerDownOutside={() => {
-                              setShare({ dismiss: "outside", open: false })
-                            }}
-                            onFocusOutside={() => {
-                              setShare({ dismiss: "outside", open: false })
-                            }}
-                            onCloseAutoFocus={(event) => {
-                              if (share.dismiss === "outside") event.preventDefault()
-                              setShare("dismiss", null)
-                            }}
-                          >
-                            <div class="flex w-full flex-col gap-1.5 px-0.5 pt-0.5">
-                              <div class="select-none text-[13px] font-[530] leading-none tracking-[-0.04px] text-v2-text-text-base [font-variation-settings:'slnt'_0]">
-                                {language.t("session.share.popover.title")}
-                              </div>
-                              <div class="select-none text-[13px] font-[440] leading-5 tracking-[-0.04px] text-v2-text-text-muted [font-variation-settings:'slnt'_0]">
-                                {shareUrl()
-                                  ? language.t("session.share.popover.description.shared")
-                                  : language.t("session.share.popover.description.unshared")}
-                              </div>
-                            </div>
-                            <div class="flex w-full flex-col gap-2">
-                              <Show
-                                when={shareUrl()}
-                                fallback={
-                                  <Button
-                                    variant="contrast"
-                                    class="w-full"
-                                    onClick={() => void props.action.share()}
-                                    disabled={props.pending.share()}
-                                  >
-                                    {props.pending.share()
-                                      ? language.t("session.share.action.publishing")
-                                      : language.t("session.share.action.publish")}
-                                  </Button>
-                                }
-                              >
-                                <div class="flex flex-col gap-2">
-                                  <div
-                                    class="flex h-8 w-full items-center gap-1.5 rounded-[6px] py-1 pl-2.5 pr-1.5 shadow-[var(--v2-elevation-button-neutral)]"
-                                    style={{
-                                      background:
-                                        "linear-gradient(180deg, var(--v2-alpha-light-2) 0%, var(--v2-alpha-light-0) 100%), var(--v2-background-bg-button-neutral)",
-                                    }}
-                                  >
-                                    <div
-                                      class="min-w-0 flex-1 truncate select-text cursor-text text-[13px] font-[440] leading-5 tracking-[-0.04px] text-v2-text-text-base [font-variation-settings:'slnt'_0]"
-                                      onClick={selectShareUrlText}
-                                    >
-                                      {shareUrl()}
-                                    </div>
-                                    <IconButton
-                                      type="button"
-                                      size="small"
-                                      variant="ghost-muted"
-                                      icon={<Icon name="outline-copy" />}
-                                      aria-label={language.t("session.share.copy.copyLink")}
-                                      onClick={() => void props.action.copyShareUrl()}
-                                    />
-                                    <IconButton
-                                      type="button"
-                                      size="small"
-                                      variant="ghost-muted"
-                                      icon={<Icon name="outline-square-arrow" />}
-                                      aria-label={language.t("session.share.action.view")}
-                                      onClick={props.action.viewShare}
-                                      disabled={props.pending.unshare()}
-                                    />
-                                  </div>
-                                  <div class="flex w-full">
-                                    <Button
-                                      variant="outline"
-                                      class="w-full"
-                                      onClick={() => void props.action.unshare()}
-                                      disabled={props.pending.unshare()}
-                                    >
-                                      {props.pending.unshare()
-                                        ? language.t("session.share.action.unpublishing")
-                                        : language.t("session.share.action.unpublish")}
-                                    </Button>
-                                  </div>
-                                </div>
-                              </Show>
-                            </div>
+                          <KobaltePopover.Content class="z-50 border-0 bg-transparent p-0 outline-none">
+                            <SessionSummaryPanel
+                              project={project()}
+                              directory={sessionDirectory()}
+                              local={!workspaceSession()}
+                              branch={data.location.vcs.info({ directory: sdk().directory })?.branch.current}
+                              baseBranch={data.location.vcs.info({ directory: project().worktree })?.branch.current}
+                              diffs={sessionDiffs()}
+                              sessionID={id}
+                              moveEligible={props.workspaceMoveEligible}
+                              moveDismissed={workspaceSuggestionDismissed()}
+                              onMoveDismiss={() => setWorkspaceSuggestionDismissed(true)}
+                              onReview={() => {
+                                setSummary(false)
+                                props.onReview()
+                              }}
+                            />
                           </KobaltePopover.Content>
                         </KobaltePopover.Portal>
                       </KobaltePopover>
-                    </Show>
-                  </div>
-                )}
-              </Show>
-            </div>
+                    )}
+                  </Show>
+                  <Show when={!parentID()}>
+                    <Menu
+                      gutter={6}
+                      placement="bottom-end"
+                      open={title.menuOpen}
+                      onOpenChange={(open) => {
+                        setTitle("menuOpen", open)
+                        if (open) return
+                      }}
+                    >
+                      <Menu.Trigger
+                        as={IconButton}
+                        icon={<Icon name="outline-dots" />}
+                        variant="ghost-muted"
+                        size="large"
+                        state={share.open || title.pendingShare ? "pressed" : undefined}
+                        aria-label={language.t("common.moreOptions")}
+                        aria-expanded={title.menuOpen || share.open || title.pendingShare}
+                        ref={(el: HTMLButtonElement) => {
+                          more = el
+                        }}
+                      />
+                      <Menu.Portal>
+                        <Menu.Content
+                          style={{ width: "120px", "min-width": "120px" }}
+                          onCloseAutoFocus={(event) => {
+                            if (title.pendingRename) {
+                              event.preventDefault()
+                              setTitle("pendingRename", false)
+                              openTitleEditor()
+                              return
+                            }
+                            if (title.pendingShare) {
+                              event.preventDefault()
+                              requestAnimationFrame(() => {
+                                setShare({ open: true, dismiss: null })
+                                setTitle("pendingShare", false)
+                              })
+                            }
+                          }}
+                        >
+                          <Menu.Item
+                            onSelect={() => {
+                              setTitle("pendingRename", true)
+                              setTitle("menuOpen", false)
+                            }}
+                          >
+                            {language.t("common.rename")}
+                          </Menu.Item>
+                          <Show when={shareEnabled()}>
+                            <Menu.Item
+                              onSelect={() => {
+                                setTitle({ pendingShare: true, menuOpen: false })
+                              }}
+                            >
+                              {language.t("session.share.action.share")}...
+                            </Menu.Item>
+                          </Show>
+                          <Menu.Item onSelect={() => void props.action.export(id)}>
+                            {language.t("common.export")}...
+                          </Menu.Item>
+                          {/* TODO: Need a V2 session archive API. */}
+                          <Menu.Separator />
+                          <Menu.Item onSelect={() => props.action.showDelete(id)}>
+                            {language.t("common.delete")}...
+                          </Menu.Item>
+                        </Menu.Content>
+                      </Menu.Portal>
+                    </Menu>
+
+                    <KobaltePopover
+                      open={share.open}
+                      anchorRef={() => more}
+                      placement="bottom-end"
+                      gutter={6}
+                      modal={false}
+                      onOpenChange={(open) => {
+                        if (open) setShare("dismiss", null)
+                        setShare("open", open)
+                      }}
+                    >
+                      <KobaltePopover.Portal>
+                        <KobaltePopover.Content
+                          data-component="popover-content"
+                          class="flex w-80 max-w-none flex-col items-start gap-3 rounded-[10px] border-0 bg-v2-background-bg-layer-01 p-3 shadow-[var(--v2-elevation-floating)]"
+                          style={{ "min-width": "320px" }}
+                          onEscapeKeyDown={(event) => {
+                            setShare({ dismiss: "escape", open: false })
+                            event.preventDefault()
+                            event.stopPropagation()
+                          }}
+                          onPointerDownOutside={() => {
+                            setShare({ dismiss: "outside", open: false })
+                          }}
+                          onFocusOutside={() => {
+                            setShare({ dismiss: "outside", open: false })
+                          }}
+                          onCloseAutoFocus={(event) => {
+                            if (share.dismiss === "outside") event.preventDefault()
+                            setShare("dismiss", null)
+                          }}
+                        >
+                          <div class="flex w-full flex-col gap-1.5 px-0.5 pt-0.5">
+                            <div class="select-none text-[13px] font-[530] leading-none tracking-[-0.04px] text-v2-text-text-base [font-variation-settings:'slnt'_0]">
+                              {language.t("session.share.popover.title")}
+                            </div>
+                            <div class="select-none text-[13px] font-[440] leading-5 tracking-[-0.04px] text-v2-text-text-muted [font-variation-settings:'slnt'_0]">
+                              {shareUrl()
+                                ? language.t("session.share.popover.description.shared")
+                                : language.t("session.share.popover.description.unshared")}
+                            </div>
+                          </div>
+                          <div class="flex w-full flex-col gap-2">
+                            <Show
+                              when={shareUrl()}
+                              fallback={
+                                <Button
+                                  variant="contrast"
+                                  class="w-full"
+                                  onClick={() => void props.action.share()}
+                                  disabled={props.pending.share()}
+                                >
+                                  {props.pending.share()
+                                    ? language.t("session.share.action.publishing")
+                                    : language.t("session.share.action.publish")}
+                                </Button>
+                              }
+                            >
+                              <div class="flex flex-col gap-2">
+                                <div
+                                  class="flex h-8 w-full items-center gap-1.5 rounded-[6px] py-1 pl-2.5 pr-1.5 shadow-[var(--v2-elevation-button-neutral)]"
+                                  style={{
+                                    background:
+                                      "linear-gradient(180deg, var(--v2-alpha-light-2) 0%, var(--v2-alpha-light-0) 100%), var(--v2-background-bg-button-neutral)",
+                                  }}
+                                >
+                                  <div
+                                    class="min-w-0 flex-1 truncate select-text cursor-text text-[13px] font-[440] leading-5 tracking-[-0.04px] text-v2-text-text-base [font-variation-settings:'slnt'_0]"
+                                    onClick={selectShareUrlText}
+                                  >
+                                    {shareUrl()}
+                                  </div>
+                                  <IconButton
+                                    type="button"
+                                    size="small"
+                                    variant="ghost-muted"
+                                    icon={<Icon name="outline-copy" />}
+                                    aria-label={language.t("session.share.copy.copyLink")}
+                                    onClick={() => void props.action.copyShareUrl()}
+                                  />
+                                  <IconButton
+                                    type="button"
+                                    size="small"
+                                    variant="ghost-muted"
+                                    icon={<Icon name="outline-square-arrow" />}
+                                    aria-label={language.t("session.share.action.view")}
+                                    onClick={props.action.viewShare}
+                                    disabled={props.pending.unshare()}
+                                  />
+                                </div>
+                                <div class="flex w-full">
+                                  <Button
+                                    variant="outline"
+                                    class="w-full"
+                                    onClick={() => void props.action.unshare()}
+                                    disabled={props.pending.unshare()}
+                                  >
+                                    {props.pending.unshare()
+                                      ? language.t("session.share.action.unpublishing")
+                                      : language.t("session.share.action.unpublish")}
+                                  </Button>
+                                </div>
+                              </div>
+                            </Show>
+                          </div>
+                        </KobaltePopover.Content>
+                      </KobaltePopover.Portal>
+                    </KobaltePopover>
+                  </Show>
+                </div>
+              )}
+            </Show>
           </div>
-        </Show>
-        <div
-          data-timeline-virtual-content
-          ref={(element) => {
-            virtualContent = element
-            props.setContentRef(element)
-          }}
-          style={{
-            height: `${virtualizer.getTotalSize()}px`,
-            position: "relative",
-            width: "100%",
-          }}
-        >
-          <For each={virtualRowKeys()}>{(rowKey) => <VirtualTimelineRow rowKey={rowKey} />}</For>
-          <Show when={timelineRows().length > 0}>
-            <div
-              data-timeline-row="bottom-spacer"
-              aria-hidden="true"
-              class="h-16 absolute top-0 left-0 w-full"
-              style={{ transform: `translateY(${virtualizer.getTotalSize() - 64}px)` }}
-            />
-          </Show>
         </div>
-      </ScrollView>
-    </div>
+      }
+    />
   )
 }
