+++ packages/app/src/pages/session/timeline/virtualizer.tsx
@@ -1,0 +1,499 @@
+import { createVirtualizer, defaultRangeExtractor, elementScroll, type VirtualItem } from "@tanstack/solid-virtual"
+import { isScrollKeyTarget, scrollKey, scrollKeyOwner, ScrollView } from "@opencode-ai/ui/scroll-view"
+import { TimelineRow } from "@opencode-ai/session-ui/timeline/projection"
+import { normalizeWheelDelta, shouldMarkBoundaryGesture } from "@/pages/session/message-gesture"
+import { useLanguage } from "@/context/language"
+import {
+  createEffect,
+  createMemo,
+  createSignal,
+  For,
+  on,
+  onCleanup,
+  onMount,
+  Show,
+  type Accessor,
+  type JSX,
+} from "solid-js"
+import { createStore } from "solid-js/store"
+import type { createTimelineProjection } from "./projection"
+import { scheduleConnectedMeasure } from "./measure"
+import { observeElementOffsetReconnectAware } from "./observe-element-offset"
+import { filterVirtualIndexes } from "./virtual-items"
+
+const fallbackItemSize = 60
+const cache = new Map<string, { measurements: VirtualItem[]; toolOpen: Record<string, boolean | undefined> }>()
+
+type Projection = Pick<
+  ReturnType<typeof createTimelineProjection>,
+  "activeMessageID" | "messageLastRowIndex" | "messageRowIndex" | "rowByKey" | "rows"
+>
+
+type Input = {
+  sessionKey: Accessor<string>
+  projection: Projection
+  showHeader: Accessor<boolean>
+  shouldAnchorBottom: Accessor<boolean>
+  hasScrollGesture: Accessor<boolean>
+  scroll: Accessor<{ overflow: boolean; bottom: boolean; jump: boolean }>
+  onResumeScroll: () => void
+  setScrollRef: (element: HTMLDivElement | undefined) => void
+  setContentRef: (element: HTMLDivElement) => void
+  onScheduleScrollState: (element: HTMLDivElement) => void
+  onAutoScrollHandleScroll: () => void
+  onAutoScrollInteraction: (event: MouseEvent) => void
+  onMarkScrollGesture: (target?: EventTarget | null) => void
+  onUserScroll: () => void
+  onHistoryScroll: () => void
+  setRevealMessage?: (fn: (id: string) => void) => void
+  setScrollToEnd?: (fn: () => void) => void
+  setHistoryAnchor?: (handlers: { capture: () => void; restore: (done: boolean) => void }) => void
+}
+
+type ViewProps = {
+  header: JSX.Element
+  workspaceSession: Accessor<boolean>
+  deferred: (row: TimelineRow.TimelineRow) => boolean
+  renderRow: (row: Accessor<TimelineRow.TimelineRow>, onSizeChange?: () => void) => JSX.Element
+}
+
+export function createTimelineVirtualizer(input: Input) {
+  const language = useLanguage()
+  const ownerSessionKey = input.sessionKey()
+  const cached = cache.get(ownerSessionKey)
+  const initialMeasurements = cached?.measurements
+  const coldBottomMount = !initialMeasurements?.length && input.shouldAnchorBottom()
+  const [listRoot, setListRoot] = createSignal<HTMLDivElement>()
+  const [toolOpen, setToolOpen] = createStore<Record<string, boolean | undefined>>(cached?.toolOpen ?? {})
+  const [renderOverscan, setRenderOverscan] = createSignal(initialMeasurements?.length || coldBottomMount ? 6 : 20)
+  const rows = input.projection.rows
+  const rowByKey = input.projection.rowByKey
+  let touchGesture: number | undefined
+  let prependAnchor: { key: string; offset: number } | undefined
+  let prependAnchorFrame: number | undefined
+  let prependLoading = false
+  let resizePinnedIndexes: number[] = []
+  let resizePinFrame: number | undefined
+  let virtualContent: HTMLDivElement | undefined
+
+  const clearPrependAnchor = () => {
+    prependLoading = false
+    prependAnchor = undefined
+    if (prependAnchorFrame === undefined) return
+    cancelAnimationFrame(prependAnchorFrame)
+    prependAnchorFrame = undefined
+  }
+
+  const capturePrependAnchor = () => {
+    prependLoading = true
+    updatePrependAnchor()
+  }
+
+  const updatePrependAnchor = () => {
+    const root = listRoot()
+    if (!root) return
+    const view = root.getBoundingClientRect()
+    const item = [...root.querySelectorAll<HTMLElement>("[data-timeline-key]")]
+      .map((element) => ({ element, rect: element.getBoundingClientRect() }))
+      .filter((entry) => entry.rect.bottom > view.top)
+      .sort((a, b) => a.rect.top - b.rect.top)[0]
+    if (!item) return
+    prependAnchor = { key: item.element.dataset.timelineKey ?? "", offset: item.rect.top - view.top }
+  }
+
+  const restorePrependAnchor = (done: boolean) => {
+    prependLoading = !done
+    applyPrependAnchor()
+  }
+
+  const applyPrependAnchor = () => {
+    const anchor = prependAnchor
+    const root = listRoot()
+    if (!anchor || !root) {
+      if (!prependLoading) clearPrependAnchor()
+      return
+    }
+    if (prependAnchorFrame !== undefined) cancelAnimationFrame(prependAnchorFrame)
+    prependAnchorFrame = requestAnimationFrame(() => {
+      prependAnchorFrame = undefined
+      const element = [...root.querySelectorAll<HTMLElement>("[data-timeline-key]")].find(
+        (item) => item.dataset.timelineKey === anchor.key,
+      )
+      if (!element) {
+        if (!prependLoading) clearPrependAnchor()
+        return
+      }
+      const view = root.getBoundingClientRect()
+      root.scrollBy({ top: element.getBoundingClientRect().top - view.top - anchor.offset })
+      if (!prependLoading) clearPrependAnchor()
+    })
+  }
+
+  const virtualizer = createVirtualizer<HTMLDivElement, HTMLDivElement>({
+    get count() {
+      return rows().length
+    },
+    getScrollElement: () => listRoot() ?? null,
+    observeElementOffset: observeElementOffsetReconnectAware,
+    initialOffset: () => (input.shouldAnchorBottom() ? Number.MAX_SAFE_INTEGER : 0),
+    initialMeasurementsCache: initialMeasurements,
+    estimateSize: () => fallbackItemSize,
+    scrollToFn: (offset, options, instance) => {
+      if (virtualContent) virtualContent.style.height = `${instance.getTotalSize()}px`
+      elementScroll(offset, options, instance)
+    },
+    get getItemKey() {
+      const items = rows()
+      return (index: number) => {
+        const row = items[index]
+        if (!row) return `removed:${index}`
+        return TimelineRow.key(row)
+      }
+    },
+    anchorTo: "end",
+    followOnAppend: true,
+    scrollEndThreshold: 80,
+    get scrollMargin() {
+      return input.showHeader() ? 64 : 0
+    },
+    overscan: 50,
+    paddingEnd: 64,
+    rangeExtractor: (range) => {
+      const id = input.projection.activeMessageID()
+      const active = id ? (input.projection.messageLastRowIndex().get(id) ?? -1) : -1
+      const indexes = defaultRangeExtractor({ ...range, overscan: renderOverscan() })
+      return filterVirtualIndexes(
+        [...new Set([...resizePinnedIndexes, ...indexes, ...(active < 0 ? [] : [active])])].sort((a, b) => a - b),
+        range.count,
+      )
+    },
+  })
+  const resizeItem = virtualizer.resizeItem
+  let resizeAnchorScheduled = false
+  const anchorResizedBottom = () => {
+    if (resizeAnchorScheduled || input.hasScrollGesture()) return
+    resizeAnchorScheduled = true
+    queueMicrotask(() => {
+      resizeAnchorScheduled = false
+      if (!input.shouldAnchorBottom() || input.hasScrollGesture()) return
+      virtualizer.scrollToEnd()
+    })
+  }
+  virtualizer.resizeItem = (index, size) => {
+    const item = virtualizer.measurementsCache[index]
+    const previous = item ? (virtualizer.itemSizeCache.get(item.key) ?? item.size) : undefined
+    const root = listRoot()
+    if (root && previous !== undefined && Math.abs(size - previous) > root.clientHeight) {
+      const view = root.getBoundingClientRect()
+      resizePinnedIndexes = [...root.querySelectorAll<HTMLElement>("[data-index]")]
+        .filter((element) => {
+          const rect = element.getBoundingClientRect()
+          return rect.bottom > view.top && rect.top < view.bottom
+        })
+        .map((element) => Number(element.dataset.index))
+      if (resizePinFrame !== undefined) cancelAnimationFrame(resizePinFrame)
+      resizePinFrame = requestAnimationFrame(() => {
+        resizePinFrame = requestAnimationFrame(() => {
+          resizePinFrame = undefined
+          resizePinnedIndexes = []
+        })
+      })
+    }
+    resizeItem(index, size)
+    if (root && input.shouldAnchorBottom()) anchorResizedBottom()
+  }
+  virtualizer.shouldAdjustScrollPositionOnItemSizeChange = (item) => {
+    if (input.shouldAnchorBottom()) return false
+    const first = virtualizer.range?.startIndex
+    return first !== undefined && item.index < first
+  }
+  const virtualItemByKey = createMemo(
+    () => new Map(virtualizer.getVirtualItems().map((item) => [item.key, item] as const)),
+  )
+  const virtualRowKeys = createMemo(() => virtualizer.getVirtualItems().map((item) => String(item.key)))
+
+  createEffect(() => {
+    input.setRevealMessage?.((id) => {
+      const index = input.projection.messageRowIndex().get(id)
+      if (index === undefined) return
+      virtualizer.scrollToIndex(index, { align: "center" })
+    })
+    input.setScrollToEnd?.(() => virtualizer.scrollToEnd())
+    input.setHistoryAnchor?.({ capture: capturePrependAnchor, restore: restorePrependAnchor })
+  })
+
+  let overscanFrame: number | undefined
+  onMount(() => {
+    overscanFrame = requestAnimationFrame(() => {
+      if (input.shouldAnchorBottom()) virtualizer.scrollToEnd()
+      overscanFrame = requestAnimationFrame(() => {
+        overscanFrame = undefined
+        if (renderOverscan() < 20) setRenderOverscan(20)
+        if (input.shouldAnchorBottom()) virtualizer.scrollToEnd()
+      })
+    })
+  })
+
+  const maybeAnchorBottom = () => {
+    if (rows().length === 0) return
+    if (!input.shouldAnchorBottom() || input.hasScrollGesture()) return
+    if (resizePinFrame !== undefined) cancelAnimationFrame(resizePinFrame)
+    clearPrependAnchor()
+    if (prependAnchorFrame !== undefined) cancelAnimationFrame(prependAnchorFrame)
+    virtualizer.scrollToEnd()
+  }
+
+  let measuredSessionKey = input.sessionKey()
+  createEffect(() => {
+    const key = input.sessionKey()
+    rows().length
+    if (measuredSessionKey !== key) {
+      measuredSessionKey = key
+      virtualizer.measure()
+    }
+    maybeAnchorBottom()
+  })
+
+  const bindListRoot = (root: HTMLDivElement) => {
+    if (root === listRoot()) return
+    setListRoot(root)
+    input.setScrollRef(root)
+  }
+
+  const handleListWheel = (event: WheelEvent & { currentTarget: HTMLDivElement }) => {
+    if (!prependLoading) clearPrependAnchor()
+    const root = event.currentTarget
+    const delta = normalizeWheelDelta({
+      deltaY: event.deltaY,
+      deltaMode: event.deltaMode,
+      rootHeight: root.clientHeight,
+    })
+    if (!delta) return
+    markBoundaryGesture({ root, target: event.target, delta, onMarkScrollGesture: input.onMarkScrollGesture })
+  }
+
+  const handleListTouchStart = (event: TouchEvent) => {
+    if (!prependLoading) clearPrependAnchor()
+    touchGesture = event.touches[0]?.clientY
+  }
+
+  const handleListTouchMove = (event: TouchEvent & { currentTarget: HTMLDivElement }) => {
+    const next = event.touches[0]?.clientY
+    const previous = touchGesture
+    touchGesture = next
+    if (next === undefined || previous === undefined) return
+    const delta = previous - next
+    if (!delta) return
+    markBoundaryGesture({
+      root: event.currentTarget,
+      target: event.target,
+      delta,
+      onMarkScrollGesture: input.onMarkScrollGesture,
+    })
+  }
+
+  const handleListTouchEnd = () => {
+    touchGesture = undefined
+  }
+
+  const handleListPointerDown = (event: PointerEvent & { currentTarget: HTMLDivElement }) => {
+    if (!prependLoading) clearPrependAnchor()
+    input.onMarkScrollGesture(event.target)
+  }
+
+  const handleListPointerMove = (event: PointerEvent) => {
+    if (event.buttons !== 1) return
+    input.onMarkScrollGesture(event.target)
+  }
+
+  const handleListKeyDown = (event: KeyboardEvent & { currentTarget: HTMLDivElement }) => {
+    const key = scrollKey(event)
+    if (!key) return
+    if (!isScrollKeyTarget(event.target, key)) return
+    if (scrollKeyOwner(event.currentTarget, event.target, key) !== event.currentTarget) return
+    if (!prependLoading) clearPrependAnchor()
+    input.onMarkScrollGesture(event.currentTarget)
+  }
+
+  const handleListScroll = (event: Event & { currentTarget: HTMLDivElement }) => {
+    if (prependLoading) updatePrependAnchor()
+    input.onScheduleScrollState(event.currentTarget)
+    input.onHistoryScroll()
+    if (!input.hasScrollGesture()) return
+    input.onUserScroll()
+    input.onAutoScrollHandleScroll()
+    input.onMarkScrollGesture(event.currentTarget)
+  }
+
+  function View(props: ViewProps) {
+    function VirtualRow(rowProps: { rowKey: string }) {
+      let element: HTMLDivElement
+      const initialItem = virtualItemByKey().get(rowProps.rowKey)!
+      const initialRow = rowByKey().get(rowProps.rowKey)!
+      const item = createMemo(() => virtualItemByKey().get(rowProps.rowKey) ?? initialItem)
+      const row = createMemo(() => rowByKey().get(rowProps.rowKey) ?? rows()[item().index] ?? initialRow)
+      const [ready, setReady] = createSignal(initialItem.size <= fallbackItemSize || !props.deferred(initialRow))
+      let contentMeasureFrame: number | undefined
+
+      onMount(() => virtualizer.measureElement(element))
+      createEffect(
+        on(
+          () => item().index,
+          () => {
+            virtualizer.measureElement(element)
+          },
+          { defer: true },
+        ),
+      )
+      onCleanup(() => {
+        if (contentMeasureFrame !== undefined) cancelAnimationFrame(contentMeasureFrame)
+        queueMicrotask(() => virtualizer.measureElement(null))
+      })
+
+      return (
+        <div
+          data-timeline-key={rowProps.rowKey}
+          style={{
+            position: "absolute",
+            top: `${item().start - (input.showHeader() ? 64 : 0)}px`,
+            left: "0",
+            width: "100%",
+            height: `${item().size}px`,
+            overflow: "clip",
+            "overflow-clip-margin": row()._tag === "TurnGap" ? undefined : "0.5px",
+          }}
+        >
+          <div
+            ref={(value) => {
+              element = value
+            }}
+            data-index={item().index}
+            style={{ "min-height": ready() ? undefined : `${initialItem.size}px` }}
+          >
+            {props.renderRow(row, () => {
+              setReady(true)
+              if (contentMeasureFrame !== undefined) cancelAnimationFrame(contentMeasureFrame)
+              contentMeasureFrame = scheduleConnectedMeasure(element, virtualizer.measureElement)
+            })}
+          </div>
+        </div>
+      )
+    }
+
+    return (
+      <div class="relative w-full h-full min-w-0" data-workspace-session={props.workspaceSession() ? "" : undefined}>
+        <div
+          class="absolute left-1/2 -translate-x-1/2 z-[60] pointer-events-none transition-all duration-200 ease-out"
+          classList={{
+            "bottom-8": true,
+            "opacity-100 translate-y-0 scale-100": input.scroll().overflow && input.scroll().jump,
+            "opacity-0 translate-y-2 pointer-events-none": !input.scroll().overflow || !input.scroll().jump,
+            "scale-[0.8]": !input.scroll().overflow || !input.scroll().jump,
+          }}
+        >
+          <button
+            type="button"
+            aria-label={language.t("session.messages.jumpToLatest")}
+            class="pointer-events-auto flex items-center justify-center w-8 h-7 px-2 py-1.5 rounded-lg border-none cursor-pointer text-v2-text-text-base backdrop-blur-[2px]"
+            style={{
+              background: "color-mix(in srgb, var(--v2-background-bg-base) 92%, transparent)",
+              "box-shadow": "var(--v2-elevation-raised), 0px 2px 8px var(--v2-background-bg-base)",
+            }}
+            onClick={input.onResumeScroll}
+          >
+            <svg width="16" height="16" viewBox="0 0 16 16" fill="none" aria-hidden="true">
+              <path
+                d="M12.3333 8.66665L8 13L3.66667 8.66665M8 12.6667V2.83332"
+                stroke="currentColor"
+                stroke-linecap="square"
+              />
+            </svg>
+          </button>
+        </div>
+        <ScrollView
+          viewportRef={bindListRoot}
+          onWheel={handleListWheel}
+          onTouchStart={handleListTouchStart}
+          onTouchMove={handleListTouchMove}
+          onTouchEnd={handleListTouchEnd}
+          onTouchCancel={handleListTouchEnd}
+          onPointerDown={handleListPointerDown}
+          onPointerMove={handleListPointerMove}
+          onKeyDown={handleListKeyDown}
+          onScroll={handleListScroll}
+          onClick={input.onAutoScrollInteraction}
+          class="relative min-w-0 w-full h-full"
+          style={{ "--sticky-accordion-top": input.showHeader() ? "48px" : "0px" }}
+        >
+          <Show when={input.showHeader()}>{props.header}</Show>
+          <div
+            data-timeline-virtual-content
+            ref={(element) => {
+              virtualContent = element
+              input.setContentRef(element)
+            }}
+            style={{ height: `${virtualizer.getTotalSize()}px`, position: "relative", width: "100%" }}
+          >
+            <For each={virtualRowKeys()}>{(rowKey) => <VirtualRow rowKey={rowKey} />}</For>
+            <Show when={rows().length > 0}>
+              <div
+                data-timeline-row="bottom-spacer"
+                aria-hidden="true"
+                class="h-16 absolute top-0 left-0 w-full"
+                style={{ transform: `translateY(${virtualizer.getTotalSize() - 64}px)` }}
+              />
+            </Show>
+          </div>
+        </ScrollView>
+      </div>
+    )
+  }
+
+  onCleanup(() => {
+    clearPrependAnchor()
+    cache.delete(ownerSessionKey)
+    cache.set(ownerSessionKey, { measurements: virtualizer.takeSnapshot(), toolOpen: { ...toolOpen } })
+    while (cache.size > 16) cache.delete(cache.keys().next().value!)
+    if (resizePinFrame !== undefined) cancelAnimationFrame(resizePinFrame)
+    if (overscanFrame !== undefined) cancelAnimationFrame(overscanFrame)
+    input.setScrollRef(undefined)
+    input.setRevealMessage?.(() => {})
+    input.setScrollToEnd?.(() => {})
+    input.setHistoryAnchor?.({ capture: () => {}, restore: () => {} })
+  })
+
+  return {
+    disclosure: {
+      value: (key: string) => toolOpen[key],
+      set: (key: string, open: boolean) => setToolOpen(key, open),
+    },
+    View,
+  }
+}
+
+function boundaryTarget(root: HTMLElement, target: EventTarget | null) {
+  const current = target instanceof Element ? target : undefined
+  const nested = current?.closest("[data-scrollable]")
+  if (!(nested instanceof HTMLElement) || nested === root) return undefined
+  return nested
+}
+
+function markBoundaryGesture(input: {
+  root: HTMLElement
+  target: EventTarget | null
+  delta: number
+  onMarkScrollGesture: (target?: EventTarget | null) => void
+}) {
+  const target = boundaryTarget(input.root, input.target)
+  if (
+    target &&
+    !shouldMarkBoundaryGesture({
+      delta: input.delta,
+      scrollTop: target.scrollTop,
+      scrollHeight: target.scrollHeight,
+      clientHeight: target.clientHeight,
+    })
+  )
+    return
+  input.onMarkScrollGesture(input.root)
+}
